requirements.txt
PySide6==6.2.2.1
PyYAML==6.0